requirements.txt
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
environs==7.3.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.6.0
marshmallow==3.5.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytest==5.4.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
environs==7.3.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.6.0
marshmallow==3.5.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytest==5.4.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
